Renamed solution to GoodStuff_UserApi. Updated nugets
