+++ Model/Models/IProduct.cs
@@ -1,4 +1,4 @@
-﻿namespace Model;
+﻿namespace Model.Models;
 
 public interface IProduct
 {
